Restrict documents pattern to those with a program component
This results in 'gsoc/document/user/foo/bar' now returning a 404.
Previously the page would render but the most template elements (such
as the sidebar and the footer) would fail to render because when we
need a GSoC program to properly render the GSoC templates, and there
is no way to retrieve one.

This should be the last page that does not specify both a sponsor
and a program, meaning that we can remove the 'Site.active_program'
bit from request_data, and use active_program only for the '/' url.