Remove use of set() to make uniqified lists.
a_set != a_list even if they consist of same elements. We can use
list(set(a)) to create a uniqified list. But, for the use case here
only one element is suspicious to be already in list. So we can
check for that element only. This reduces complexity from O(n^2)
to O(n). Or can use set.union, then convert result to a list.