Delete account logic added.
At this moment it actually does not do much, but sends an email to application admins who are supposed to delete all the data on their own.
In the future, it should be an automated process. We are waiting for the legal team to give us some feedback to our questions related to how the deletion of entities should be handled.