If the entered email address is invalid show the relevant form error message and re-raise the exception.
If we don't re-raise the exception the line

account = users.User(email)

gets executed and since there was an error while cleaning email field, the
cleaner did not return anything but raised an exception, which means that
the variable email never got defined. This will cause an error of attempting
to access undefined variable. To avoid this, bail out when we know that
the email address is invalid.