Style fixes to stopwatch code
While descriptive names are important in API design (e.g., the
naming of the public functions), 'short, sweet, and to the point'
is more important when naming variables.

Long variable names are hard to read, and make it difficult to scan
the code and see at a glance what is going on. If from context it is
clear that a variable refers to the total duration in seconds (e.g.,
because it is assigned by the 'totalRemainingSeconds'), just
'remaining' is sufficient as a variable name.