As a first step of refactoring move the tags processing code from the view to the form cleaner.
This occured to me when there was an exception whenever the form errored
out because of invalid "score value" to difficulty tag and tried to raise
the ValidationError exception. This exception is caught and form errors
are set only if this exception occurs within the form validation chain.
Outside that chain, there is simply no code to catch that exception and
it bubbles all the way upto the Application level and thus causes an HTTP
500.

It is fixed now. But this tag code still needs one more round of refactoring
to move the logic specific code to tag logic.