This is the change in jquery that caused the regression in uniform plugin not to let unchecking of checkboxes.
A big thanks to that guy John on Stackoverflow. I have left the comment
and the link to stackoverflow page where he gives the solution where I
found the solution there in the code intentionally to give credit to him.
I had come up with a somewhat close solution, but it did not work for the
opposite case, that is checking the box after unchecking it. And also my
solution was fairly complicated with dirty state tracking. This solution
by John as simple and elegant as it seems is not an easy one to figure out!
And it was hard to find too.

Fixes issue 1339.