Refactor the code that fetches the work submissions to ensure a valid blob is being referenced.
This check was being done in the template previously using the .get method on the
BlobInfo instance. But .get method is really a method on the class and in addition
it also needs the BlobInfo entity key as the argument always which we cannot pass
in the Django templates while calling methods. So moved this logic to Python side
and thus build a dictionary of True-False values for validity of the Blobs. We
also pass the actual upload_of_work blob in this dictionary since we have already
fetched it once.