Use the path to the images in Melange's image path.
The old path is the one supplied by EchoDitto. We need to replace them
everytime we get new CSS.

This 2000+ lines of single CSS file (as opposed to split multiple files
for GSoC) is such a mess. It is so painful to locate things :(

Also there is no {less} so much repeatition, nothing less than PITA :(