Allow dev_appserver to use crypto C extension. Details in the issue report below.
http://code.google.com/p/googleappengine/issues/detail?id=4168
This patch was also submitted in the bug report itself. The bug
probably only hits Ubuntu/Debian users. We keep applying this every time
we update the SDK. I was hoping that this will be fixed with the SDK
that supported Python 2.7 but looks like 1.6.0 doesn't fix this too.
Appengine fail :)