Needs the key for each list item to be set explicitly based on the student profile entity key.
This is because the default key which is the key for the GCIScore model
is same for every list item since there is only one such entity per
program and hence all lead to same key. This causes a problem while
generating links for row actions. To make each list item uniquely
identifiable we assign the key of the student profile entity.