Do the initialization of the task_type field when there is an instance bound to the form.
The other cases of initialization like initialization when the previous
form submission failed with values on this field is automatically taken
care by Django. Django cannot take care of instance based initialization
because task_type is not a regular Appengine property. So these values
have to be fetched by us.