Links on comments by students leads the list of their tasks.
The solution can be improved - right now, we take a User and check if he or she is a student by fetching their profile. It requires running a query.
Comments should either have a field to determine if they have been posted by students or point to the profiles.

Fixes issue 1356.