A script to download GCI student forms added.
The script could be more efficient, if we had an easy way to determine if a student has completed at least one task based on GCIProfile or GCIStudentInfo. Without it, we download GCIScore first.

It also takes very long to prefetch all the GCIProfile entities which are parents of the score entities. Maybe it is worth to consider just deepFetching all profiles for the program and filter out the ones we do not need.