Expect a JSON associative array containing the blobstore upload link instead of a regular string.
This is because we now use jsonContext method instead of directly
using the json method in the views to send the JSON response. The
jsonContext method can either return a string or a context dictionary.
We are choosing context dictionary to be consistent with how the
Django views work and hence how the regular context method works.