New ProgramMessages and GSoCProgramMessages models added.
Their purpose it to store program wide messages which may be sent to participantin emails at various points of the program.
The functionality to actually send emails must be added separatetly for each property in this module.
The entities of this model should be descendants of Program entities. There should be only one entity per program.
This fields are not kept in Program instance, because they are berely accessed (usually a few times throughout the program), while their presence could increase the size of program entities. We do not want to do that, because program entities are accessed very frequently.