Change the logic that disables buttons in form when submitting.
We couldn't disable the buttons, otherwise the name of the button
is not communicated to the backend, hence failing in forms with multiple
form buttons. We just prevent the form to be submitted twice using a
fake class and adding a disabled class to all the buttons. This is
rolled back after 3 seconds to provide a fallback.