Return the user entities for the host and not the host entities for now.
This is because we are not storing the host entities along with
their corresponding user entities, so there is no way we can
get to the user entities from the host entities and vice-versa.
So for now let us use the email IDs from the user entities of
the host. However, we should convert all the host entities to
have the corresponding user entity as the reference.