Check if the current time is in between the start and end dates on the timeline.
Currently we just check if time now is after certain start date. Because
of the way if conditions are layed out, the previous condition is always
true when the next condition is true. That is when the students are announced,
afterStudentSignup is always true, so the check returns the time period at
student signup point already and not students announced.

Fixes issue 1535.