Fix issue 1517: Profie Page does not show the role of the user
There were two solutions to fix this. One was to include 'is_mentor',
'is_student' and 'is_org_admin' in GSoCProfileReadOnlyTemplate and this would
end up in three rows displaying the boolean values in the web page. I
believe only displaying the current role should be enough. So, in this patch
I display the appropriate role as a big heading in the top right corner of the
page.