Refactor prefetcher to run db.get asynchronously.
In order to not break the existing code, use the asynchronous
prefetcher methods to perform synchronous prefetching by
waiting for the results and then performing the post processing
tasks on the fetched entities.

On the other hand the code that uses the asynchronous versions
of the prefetcher must do the post-processing on the entities
themselves, otherwise what's the point in prefetching asynchronously? :)