Students should be able to upload the forms even before they complete a task.
This is the new rule for GCI and so we show the link to form
upload page in the student dashboard as soon as they create
a student profile.

Further we show the warning message on the top of the student
dashboard when one of the forms is not uploaded but we don't
show the link to the form upload page anymore since that makes
the link to the form upload page redundant. The link is already
there in the dashboard.