For participating organization logos use "title" as well as "alt".
It turns out that modern guidance is to use alt to describe an image
and to use title for the title of the thing the image represents, so
our alt properties are now "<organization name> logo" and our titles
are now "<organization name>".

This has the benefit of affording mouseover tooltips on participating
organization logos!