Use the template file for the field instead of the hard coded string.
We render this template to string by using the same dictionary we used to
populate the string as a context dictionary.

Also, since django's form util flat_attr method was really written to be used
within the Python code and not for templates, it double quotes the already
quoted strings for HTML attribute values when used in conjunction with
render_to_string, so we re-use the necessary part of that method within this
method.