Implement both the read only profile show page for the users and the program hosts.
The regular users version and the program host version are entirely
two different pages with two different URLs. This was necessary because
program host version contains certain administrative elements like
being able to view the forms that students upload and to verify them
and to keep things clean, it is a good idea to implement them as two
separate views.