testPostButtonUnpublishReopenedTaskForbidden added.
It is a simple test case which checks that trying to send a POST request to unpublish a task whose status is Reopened is forbidden. It is important to check it, because if org admins were able to unpublish such tasks, they could also edit and publish them again. It would be a violation against the policy that a task which has been claimed once is not editable.