testPostSendForReviewClosedTaskForbidden added.
This test case checks if it is possible for a student to send their task for review, if its status is already closed. Unfortunately, this test fails at this moment which is a problem if someone sends such a POST request. Our logic will change the status back to NeedsReview, although a closed task should be immutable.