It is checked if a task's status when a student send it for review.
It should not be possible to send a closed task for review, because it will corrupt internal data by setting task's status back to NeedsReview. This check here also fixes the test case added in the previous commit.

Another change in this review is a modification of an error message which is shown to the user. Adding the information about closed tasks makes it more verbose to the user.