Kill some horsehockey in all three implementations of RequestHandler.error.
The implementations of RequestHandler.error take an optional "message"
parameter, and if that parameter is not supplied, they will use the
"standard" messages for the given status code (their other parameter). To
do this, they were making use of the Response object's set_status method,
which performed the translation from status code to "standard" message.

Unfortunately, Response.set_status also *set* the Response object's
"content" attribute to be that message. To make use of the message, the
implementations of RequestHandler.error were accessing it and storing it
in their context dictionaries. They then had to clear (set to be the
empty string) the Response object's content attribute so that their
subsequent render() call (which made use of the context dictionary
containing the message) could wholly set the Response object's content
attribute (because the Response object's content attribute is actually
a property that when assigned to actually takes on the value of what was
assigned to it appended on to what was at it before).

The solution implemented here (unfortunately thricefold, but that's
another cleanup for another time): just have the implementations of
RequestHandler.error do their own looking-up of what the "standard"
message for a given HTTP status code is.