Use django.http.HttpResponse rather than soc.views.helper.response.Response.
The set_status method on soc.views.helper.response.Response was being used
here to overwrite the status code and content of the self.response object,
setting them to 200 and the empty string. Those are the default values of a
new django.http.HttpResponse object, so just return that instead.

This change was discussed in the review thread of commit e3953726fd8f3e111.