site is now a property, not an object attribute.
It changes the way how it will be accessed from now on. It initializes the _site field on demand. Therefore, it does not need to set it in populate() function anymore. From client's perspective it does not really matter, as they may access this field in a regular way: by calling request_data.site