POST dictionary is not copied in _getCreateProfileForm.
I do not understand the reason why it was copied in the first place. It may have caused confussion though.
So what we did was to first copy an object under field POST and then reassigned that new object to the same field again. In other words
  self.foo = self.foo.copy()
When could it make sense? One example would be, if an object stored under self.foo is shared with other objects and we want to prevent local changes from being visible in other places.
Well, the only field, which seems to reference the same object as self.data.POST, is self.data.request.POST. Actually, this self.data.POST is a shortcut for that field, so I would expect that they both point at the same object.