organization field is initialized on demand.
I have two serious concerns about this field. Firstly, I am not sure if it is
the right place to store it at this moment. Basically, it retrieves an entity
based on optional kwargs in request. I would rather see it in Mutator as
orgFromKwargs. Secondly, this value is overriden in some other places. IMHO it
is dangerous and seems like asking for trouble, because we reuse the same (more
or less) global variable for different purposes. It is only the programmer who
knows what it refers to depending on the view. I did not change these all values
right now, but added a TODO instead. Also, this is the reason why this property
has a setter.