redirect field is now a property in RequestData object.
It makes it easier to manage the
two objects. Before this change, we first had to create an "unpopulated"
RequestData object, then use its reference to construct RedirectHelper and
finally we were able to populate RequestData object with a reference to
RedirectHelper. It also created an circular references. Right now, it is all
simplified.