Sanitize the input for the free-form text fields (input types text and textarea).
This sanitization escapes the script tags and any HTML code.

This should in fact be implemented in the base class that generates
the char field, but that means that we need to copy/paste a lot of
code from Appengine's modelform module in order to monkey patch this
code. But we have plans of decoupling Melange's forms from that module
because of the planned changes w.r.t. datastore layer, so I felt it
is not worth to take that route and followed a simpler approach of
explicitly calling the cleaner on the free-form fields for now.