Another approach to handle different types in different list columns.
Different types will be defined in different classes which inherits from the abstract ColumnType class. This class defines an abstract safe() method whose objective is to take the specified value and return a safe representation which may be put as HTML on the page.
Currently I have added three different types: plain text, numerical and HTML.
The first one may contain arbitrary values. The raw form may be vulnerable to various malicious inputs. Therefore, the output is HTML escaped.
The second one may contain numbers. These values are not escaped so as to be ordered correctly by JS Grid.
The last type may contain HTML content. As I mentioned above such input may be dangerous, so by declaring this type a developer confirms that he or she is aware of that. This will be rendered as HTML on the page.

Also, a factory class is added. It has create() method which returns an instance of the type which corresponds to its string representation. The valid string representations are defined in ColumnType class, but it is a subject to change.

Another change is that methods to create columns with specific types are added. Specifically, those are addNumericalColumn and addHtmlColumn.

Last but not least, explict 'escape' parameter is removed. The reason is that I am not sure if it is safe to exposed directly to the clients. It puts another burden on developers, as they may accidently define the configuration in a wrong way. For instance, passing escape=False to columns which may contain user's input is not safe. Similarly, not passing escape=False to columns which contain numbers will result in a wrong order. IMHO the two added functions are more clear.