Functions creating specific column types are used in preference to addColumn.
This is a follow-up to the last number of changes regarding the columns. It is better if a client must explictly say what type of column is required.
Also, there were some errors which resulted from the fact that addColumn added a PLAIN_TEXT column by default. The content was HTML escaped which was not always desired.

There is probably still a couple of errors with rendering which we will be caught sooner or later. This revision fixes at least 10 of them.