Set the available "Active Program" values in EditSitePage.context.
They were being set at SiteForm.Meta class definition time (despite
the attempted lazy evaluation of getProgramMap) due to the fact that
django_widgets.Select evaluates and copies its "choices" argument
at instantiation.

This problem was being masked by the way that until revision
09cae9038b8ea23b9b95b31e3ce7ca6fde48e195 it just happened to be the
case that soc.views.site was being loaded into the Python
interpreter after the calls to core.registerModuleCallbacks that
populate the global list of available programs.

This code is a workaround; the long-term solution is for the list
of available programs to be a request-time function of the data
store rather than an interpreter-instantiation-loaded, never-updated
global field.