full_name property added to Profile model.
It returns the name in the following format: 'first_name last_name'. We generally do not want to use it, as public_name or name_on_documents is preferred, but in some cases it may be required to use this format. An example is a list of GCI Grand Prize Winners.