winner_for property added to GCIStudentInfo model.
It may contain a reference to GCIOrganization entity for which the user was chosen as a Grand Prize Winner.

What is important, there is no reference_class argument set for the property. The problem is that GAE does not really handle circular references very well and there is already a reference to GCIProfile in GCIOrganization. As a result, there will be no runtime check if the property actually refers to GCIOrganization. It may be improved if we ever decide to switch to NDB which addresses the problem.