Extra links removed, if a person is not allowed to leave a comment.
This section made little sense.
First of all, if a user does not have a profile for the program, it will suggest student registration. What if a person is a prospective mentors? He may end up registering as a student and a developer intervention will be required.
Secondly, if a person is not logged in, he or she will be shown a login link. It is fine, but it suggests that a person will be eligible to post a comment afterwards. It is not true.
Finally, this works weried if a person is logged in with many accounts.
This issue was also discussed with Carol Smith who agrees that we should enitrely remove this link.
Fixes issue 1709