edit_gsoc_program URL pattern is moved to another pattern.
Before this change it was assigned to 'gsoc/program/sponsor_id/program_id' pattern. After this change it will refer to 'gsoc/program/edit/sponsor_id/program_id' pattern. The reason for this revision is that the former pattern is not clear enough and does not suggest that actual editing is involved.