GSoCCreateProgramPage added to program view in GSoC module.
It will allow Program Administrators to create new GSoC programs without any developer intervention. The process is that they just need to fill out the information in CreateProgramForm. There is currently one flaw: some important fields (documents, to be more specific) had to be removed, although they do exist in the edit program form. The problem is that these documents are identified to key names which depend on the program key name. When the program is created, it is not known what its link_id will be.

Currently, a Program Administrator has to create a new program without setting the documents. He or she may do it afterwards, when the program is persisted and the edit form is available. I am not sure if this is the best user experience. It is not obvious for the end user why the documents are not present on the initial form.

If we want to fix it, I think we can separate the documents from the Program model or at least remove them from the edit program model and put them in a different form. It would be consistent with some existing solutions, for example with setting program messages.