djangoforms copied from AppEngine to app directory.
This is a next step towards migration to Python 2.7 SDK. djangoforms are not supported in that environment, so we need a separate copy in order to be able to continue using it. In the future we will drop this module permanently, as we are likely to separate forms from database models.

This is discussed in issue 1725.