testCreateDocument tests fixed for GSoC and GCI.
There was a new dashboard_visibility field added to Document forms in revision ad276e8b936c and revision 5a96a9033c81.

Here is an interesting part. It took me a lot of time to fix this error. At first, I tried to add a single value as visibility:

dashboard_visibility: ['student']

Unfortunately, the tests still did not pass and the form field showed the following error message: "Enter a list of values". I investigated a little bit what was going on. Firstly, I opened a real page in my browser, filled out the form and checked with a debugger what was the value in POST the POST dict. It was ['student'] as expected. On the contrary, when I was running the tests, the printed value was 'student'- not wrapped in a list anymore. That was weird. I looked a little bit further and apparently by calling response = self.client.post(url, postdata) in DjangoTestCase's post() the value is magically changed at some point. Unluckily, we do not have much control over it :(