Program logic module created and getSponsorKey function added.
This utility function will be used by clients who want to know what is the key of the sponsor for the specified program. Typical usage wil be:

program_logic.getSponsorKey(program).name()

in order to obtain sponsor's name. Previosly, that value could have been retrieved by accessing scope_path property of Program entities. The will not be used in the future. The imlementation is quite simple and straightforward, but it uses get_value_for_datastore on scope property and is long, so it deserves its own function. Also, currently it queries scope property which is likely to be removed as well. It will be easier to adopt that change if it is a separate function.