Two different views which are about project updating use canUpdateProject to check access.
The logic also apply, so the code may be reused. This basically extends the range of users who are able to upload the projects and upload code samples. This is related to issue 1586 which is supposed to allow Organization Administrators and Program Hosts to upload code samples on students' behalf.