_isUpdateLinkVisible function added.
It answers the question whether the current user should be displayed "Update" project link. Previosly the logic was much simpler as only the student who owned the project was supposed to see the link. Currently, this singleton group is extended with certain organization administrators and program hosts. The logic became more complicated and deserves its own function now.