OrgsListPage renamed to AdminOrgListPage and moved to accepted_orgs modules.
There is a couple of reasons for doing that. First of all there is a number of views/request handlers in admin modules. This is not entirely convenient, because it makes an impression they are thigtly bound to the Admin Dashboard. This is partially true: these handlers can currenly be accessed only from the dashboard, but on the other hand they should be considered standalone views. In other words, they may exist without the admin dashboard. Because there is a lot of such handlers in the admin modules, it makes its code very long. Secondly, all views/handlers related to accepted organizations should be put in once place. It turns out that there is a large number of similar classes and their functionality is redundant. It would be easier to spot such situations if classes are placed in a reasonable way.