Teach addKeyColumn to automatically include parent keys
Currently, addKeyColumn only includes the id of the entity. As a result,
entities that have a parent key could not use the default method, but
would have to override it with their own implementation. Otherwise,
entities with the same key (but different parent keys), would have the
same value in the key column.

This greatly simplifies our list code, as we no longer need to manually
specify the key column for the many entities that have a parent key.