A unversal class which may be used to obtain a template with list of all organizations added.
This is an initial version which will be changed. Currently it is more or less a copy of a template from accepted_orgs view. The reason for having this template in a separate view is that a list of organizations is needed in a lot of places and a lot of code is duplicated.