PROPOSAL pattern is extended.
Previosly it was sponsor-id/program-id/proposal-id. The last element (proposal-id) was not enough to fetch a GSoCProposal, as this model has a parent which points to a student who is the author. Therefore, the appropriate GSoCProfile entity was needed. The code assumed that PREVIEW pattern would be used only in the context of such a student. If the author tried to access a page with PREVIEW pattern, the parent could have been retrieved from RequestData. In other words, GSoCProposal could have been obtained in the following way:
    data.proposal = GSoCProposal.get_by_id(
        int(data.kwargs['id']), parent=data.profile)
Theoretically it is all right, but its limitation is that it cannot be used for other users. There is a number of reasons why such patterns should be avoided.

This change modifies the pattern to sponsor-id/program-id/student-id/proposal-id.