Founder property removed from Group model.
It is used to describe who creates the entity. This field is filled correctly, but the information provided is completely useless. For example, we do care who is organization administrator for the organization. There is no value added by the field which states who initially created the profile. It is actually not used anywhere.