Replace "self.check" with "check" in a checkAccess implementation.
This fixes issue 1763.

Also included are several other fixes related to the RequestHandler
refactor that happened while Connection was off in a branch:
RequestHandler.post returns an HTTPResponse object rather than setting
a "response" attribute and the RedirectHelper object is an attribute
of the RequestData object now rather than the RequestHandler itself.