POST method of Profile view for GSoC always redirects to EditProfile view.
It was supposed to do something more sophisticated based on connections. The logic was so complicated that it should be rewritten if necessary.

It also threw 505 expcetion because of the line:

user = User.get_by_key_name(data.request.POST['public_name'])
data.redirect.connect_user(user=user, organization=organization)

first of all, User entities should be fetched by link_id. Here, public_name is used. This is not equivalent in majority of cases (unless a user pick the same public name as their link_id), so no entity was returned. Therefore, None was passed to connect_user, while the body of that function assumes it is not None.

It is related to issue 1778.