TODO additions and cosmetic formatting changes.
The TODOs are mostly related to violations of the modules/gsoc-melange
border. One is related to what looks like ad-hoc polymorphism ("Meh,
email-address and list-of-email-address aren't all that different!").

The cosmetic changes are (1) reordering the functions in this module
so that definitions appear before call sites, (2) stripping trailing
whitespace, and (3) spaces in the right (non-trailing) places.

This commit contains no semantic code or behavioral changes.