Author of comment is set properly to the org admin who initialized the connection.
It used to be set as the person with which the connection was initialized. The reason was that profile variable was happily used by the transactional function, because it was actually declared between the function body and the place at which it the function was called. That was the reason why NameError was not raised :-( It is described in issue 1806.