Logic for connection messages added.
Right now it contains only a single class which builds queries for GSoCConnectionMessage model. This separates actual query construction logic from views so that db related logic is not inlined in views where it does not belong to.
In some other places we try to achieve the same by adding various combinations of query names tologic, i.e. queryForPropertyAndAnohterProperty. Anyway, it generates a lot of functions which are very similar to each other. I think the approach with a builder pattern is cleaner.