hasMentorProposalAssigned function added.
As its doc string says, it can be used to determined if the specified profile (mentor profile, to be more specific) has at least one proposal for which he or she is assigned as a mentor.

Also, unit tests are added. The tested function use non-ancestor query but the test environment does not use HR schema by default.