ValueError raised by a function of URLField class caught and handled.
This was achieved by adding our own implementation of URLField class. The method in question is overriden but the only thing it does is invoking the parent function and catch possible ValueErrors.

This is related to issue 1833.

As described in doc string, this class is not needed when Django is updated to version at least 1.4.