org argument replaced by org_key in a few functions.
Previously, the functions accepted either GSoCOrganization entity or db.Key instances. Unfortunately, when anyone wanted to access organization's link_id or key_name (there are equal), isinstance had to be used to determine if the initial argument was an entity or a key.

The current solution requires function clients to construct keys on their own.

Also, unit tests are changed accordingly: keys are passed instead of entities.