Proposal view uses resubmitProposal function to resubmit a proposal.
Also, a unit test for this functionality is included.

The test is currently failing. This is because one of the logic functions (canSubmitProposal to be more specific) accesses program's timeline by calling
program.timeline
This has a hidden get() operation and introduces another roundtrip to datastore. If this is run within a transaction, it recognizes that program timeline is in different entity group than profile. Therefore, if the function is run within a transaction, then BadRequestError: cross-groups transaction need to be explicitly specified (xg=True) is raised.

In order to fix it, timeline entity will be added to canUpdateProposal function in the subsequent revisions.