timeline entity is added to canSubmitProposal function.
Therefore, it does not have to be accessed via program entity by calling program.timeline. It does not require a possible extra trip to datastore. If a get operation is required and the function is executed within a transaction, it complains and raises BadRequestError. The reason is that the get operation involves another entity group.

The further explanation is in log message to revision d0dc67b75d79c48b40146b73f2e1fd7d6185fa2c.