Add new views for server errors and user errors.
The first and most virtuous consequence of this change is the
separation of "errors" into "server errors" and "user errors".
The two represent vastly different spaces of problems and we
can speak more clearly by differentiating between the two as
much as possible. To that end: this change retains the existing
blue-on-white styling for user errors, but inverts it for an
orange-on-black styling for server errors.

This change does away with the static "404.html" and "500.html"
canned responses configured to be sent by Django and routes all
error handling through the single soc.views.helper.error._handle
function and error.html template.

This change drops an unnecessary div-containing-only-a-div in the
error.html template.