Remove an ignored, side-effect-free string calculation expression.
This has a rather cool history to it. When it was introduced in
revision 1fcb82ab (19 March 2011) it was assigned to a field that
was subsequently used. Revision 5aa9f2f6 (25 March 2011) dropped
the use of that field, but the assignment remained. Revision
c1f54d94 (20 May 2011) cleaned up the assignment to an unused
field, but left the expression itself, and it's stayed with us in
the intervening almost-two-years.

Cool stuff.