Add new methods to connection logic.
These changes were poached (via copy and paste, sadly) from a
different (no longer existing) branch in the interest of keeping
a cleaner commit history. This refactor was necessary in order to
begin moving the Connection entity logic out of the view module and
into a more accessible and reasonable location.