Add logic for connection messages and associated tests.
Incorporated Daniel's changes from revision 679056e55b6c via copy
and paste to avoid causing merge conflicts. Also made minor changes
to the test structure to make it possible to re-use utilities that
were initialized in setUp(). Structural (not logical) changes were
also necessary for the test case itself to account for the
availability of the aforementioned utilities. Made minor style
changes after discovering that the closing parenthesis on a method
call does not have to be isolated according to our style guide.