Slot usage is not checked when proposal is manually accepted.
When a program administrator manually accepted a proposal, it was checked whether the corresponding organization has not used up all their slots. This condition is removed now.

Organizations claim a slot before proposals are automatically accepted. They do that by switching 'accept_as_project' flag to True and assigning a mentor to a potential project. Slots counter is only used to make sure that an organization does not get more slots than it was initially assigned by program administrators. After the projects are set, this number does not actually keep the number of projects. I.e. it does not get updated when a project is withdrawn, fails or anything. This would actually not be trivial.

I think that 'slot claiming phase' should be distinguished from 'number of projects'. If there is a need to store the latter number, an extra counter can be added Query.count could probably be used.

In this very case, this check is not necessary, as the action can be performed only by program administrator. It can be assumed that they know what they are doing. Having used up all the slots should not prevent them.

It fixes issue 1866.